Run Partition Pipeline on Neuroimaging Data

This function initializes and executes a partitioning pipeline designed for processing neuroimaging data. It handles tasks such as image processing, super partition analysis, mapping, and combining of data based on specified thresholds and parameters.

Usage

run_partition_pipeline(
  tind,
  nfl,
  main_dir,
  tissue_type,
  ICC_thresh_vec,
  num_cores = 1,
  suppar_thresh_vec = seq(0.7, 1, 0.01),
  B = 2000,
  outp_volume = TRUE
)

Arguments

tind

Index or identifier for the type of tissue under analysis.

nfl

List of file names (full paths) that need to be processed.

main_dir

Main directory where outputs and intermediate results will be saved.

tissue_type

Type of tissue for segmentation.

ICC_thresh_vec

A vector of threshold values for Intraclass Correlation Coefficient used in the Partition Algorithm.

num_cores

Number of cores to use for parallel processing. Default to 1.

suppar_thresh_vec

Optional; a sequence of threshold values used in Super Partitioning. Default is a sequence from 0.7 to 1 by 0.01.

B

Optional; the maximum size of modules to be considered in partitioning. Default is 2000.

outp_volume

Optional; a logical indicating whether volume outputs should be generated. Default is TRUE.

Value

The function does not return a value but will output results directly to the specified main_dir as side effects of the processing steps.

Details

The function configures and runs a series of operations that are typical in neuroimage analysis, especially focusing on ROI-based transformations. Each step of the pipeline, from initial image processing (iproc) to the final combination of independent variables with reduced variables (Cmb_indep_with_dep), is executed in sequence. Adjustments to the pipeline's behavior can be made by changing the function parameters, allowing for custom analysis flows.